// src/app/models/oder.ts
export class Oder {
    "_id": string;
    "methodofpayments": string;
    "totalmoney": string;
    "items": [];
    "address": string;
    "note": string;
    "phone": string;
    "fullname": string;
    "published_at": string;
    "createdAt": string;
    "updatedAt": string;
    "id": string;
}